// src/pages/RecallProducts/types/MedicineRecallInfo.d.ts
export interface MedicineRecallInfo {
    PRDUCT:               string; // 제품명
    ENTRPS:               string; // 업체명
    RTRVL_RESN:           string; // 회수 사유
    ENFRC_YN:             string; // 강제 여부
    RTRVL_CMMND_DT:       string; // 회수 명령일
    RECALL_COMMAND_DATE:  string; // 회수 명령 날짜
    ITEM_SEQ:             string; // 품목 일련번호
}